# switch builder implementation to use chunks