--- a/small-bytearray-builder.cabal
+++ b/small-bytearray-builder.cabal
@@ -36,6 +36,7 @@ flag checked
 
 library
   exposed-modules:
+    Data.Bytes.Chunks
     Data.ByteArray.Builder
     Data.ByteArray.Builder.Unsafe
     Data.ByteArray.Builder.Bounded