--- a/small-bytearray-builder.cabal
+++ b/small-bytearray-builder.cabal
@@ -91,7 +91,7 @@ benchmark bench
     , small-bytearray-builder
     , text-short
     , byteslice
-  ghc-options: -Wall -O2
+  ghc-options: -Wall -O2 -ddump-to-file -ddump-simpl -dsuppress-all
   default-language: Haskell2010
   hs-source-dirs: bench, common
   main-is: Main.hs