Resolves https://github.com/andrewthad/small-bytearray-builder/issues/9