@andrewthad 2019-10-09
Before this PR:

```
unbounded/csv-no-escape                  mean 2.666 μs  ( +- 45.04 ns  )
unbounded/word-16-tree-small             mean 169.2 ns  ( +- 13.02 ns  )
unbounded/word-16-tree-2000              mean 29.61 μs  ( +- 624.8 ns  )
unbounded/word-16-tree-9000              mean 120.8 μs  ( +- 1.108 μs  )
```

After this PR:

```
unbounded/csv-no-escape                  mean 2.874 μs  ( +- 56.72 ns  )
unbounded/word-16-tree-small             mean 209.9 ns  ( +- 3.141 ns  )
unbounded/word-16-tree-2000              mean 33.35 μs  ( +- 925.8 ns  )
unbounded/word-16-tree-9000              mean 138.3 μs  ( +- 1.636 μs  )
```

So this currently loses performance.