@andrewthad 2019-10-09
With additional changes, performance is now:

```
unbounded/csv-no-escape                  mean 2.994 μs  ( +- 390.2 ns  )
unbounded/word-16-tree-small             mean 186.4 ns  ( +- 3.620 ns  )
unbounded/word-16-tree-2000              mean 31.86 μs  ( +- 600.6 ns  )
unbounded/word-16-tree-9000              mean 135.7 μs  ( +- 4.728 μs  )
```