@andrewthad 2019-10-09
This approach regresses performance by about 10% in the benchmarks here. I'm going to merge this anyway. In `byte-tree-2000` in `bytes-builder-shootout`, performance has not regressed. I'm still not able to complete with `fast-builder`. I suspect that this is because `fast-builder` is free to work on `Addr#` directly (pinned memory) while here we must use `MutableByteArray#`. There are places where this frees up `fast-builder` to use fewer registers since it can just keep track of the address to the current offset and the last address in the buffer. This also means that writes to memory don't have add the offset to a base address and then add 16 on top of that the way that `MutableByteArray#` requires them to. I'm not totally sure what leads to the difference. Oh well, that's a problem for another day.